requirements.txt
Flask==2.1.1
requests==2.27.1
python-dotenv==1.0.0